# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardAuthInput {
  token: string;
}

export interface DashboardAuthResult {
  ok: boolean;
}

export interface BotStats {
  tag: string;
  guildCount: number;
  userCount: number;
  uptimeSeconds: number;
  memoryMB: number;
  wsPing: number;
  online: boolean;
}

export interface GuildSummary {
  id: string;
  name: string;
  /** @nullable */
  iconURL?: string | null;
  memberCount: number;
}

export type GuildConfigSecurityLevel = typeof GuildConfigSecurityLevel[keyof typeof GuildConfigSecurityLevel];


export const GuildConfigSecurityLevel = {
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
} as const;

export type GuildConfigVpnCheckAction = typeof GuildConfigVpnCheckAction[keyof typeof GuildConfigVpnCheckAction];


export const GuildConfigVpnCheckAction = {
  kick: 'kick',
  ban: 'ban',
  flag: 'flag',
} as const;

export interface GuildConfig {
  /** @nullable */
  logChannelId: string | null;
  /** @nullable */
  banLogChannelId: string | null;
  /** @nullable */
  generalLogChannelId: string | null;
  raidMode: boolean;
  raidMode2: boolean;
  joinLock: boolean;
  /** @nullable */
  ticketStaffRoleId: string | null;
  /** @nullable */
  ticketCategoryId: string | null;
  /** @nullable */
  transcriptChannelId: string | null;
  welcomeEnabled: boolean;
  /** @nullable */
  welcomeChannelId: string | null;
  welcomeMessage: string;
  leaveEnabled: boolean;
  /** @nullable */
  leaveChannelId: string | null;
  leaveMessage: string;
  captchaEnabled: boolean;
  /** @nullable */
  captchaChannelId: string | null;
  /** @nullable */
  captchaUnverifiedRoleId: string | null;
  /** @nullable */
  captchaVerifiedRoleId: string | null;
  sanctionDmEnabled: boolean;
  /** @nullable */
  inviteLogChannelId: string | null;
  securityLevel: GuildConfigSecurityLevel;
  antiInsultEnabled: boolean;
  antiInsultWords: string[];
  antiWebhookEnabled: boolean;
  suspiciousCheckEnabled: boolean;
  whitelistedInviteCodes: string[];
  vpnCheckEnabled?: boolean;
  vpnCheckMinAgeDays?: number;
  vpnCheckAction?: GuildConfigVpnCheckAction;
  vpnCheckRequireNoAvatar?: boolean;
}

export type GuildConfigPatchSecurityLevel = typeof GuildConfigPatchSecurityLevel[keyof typeof GuildConfigPatchSecurityLevel];


export const GuildConfigPatchSecurityLevel = {
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
} as const;

export type GuildConfigPatchVpnCheckAction = typeof GuildConfigPatchVpnCheckAction[keyof typeof GuildConfigPatchVpnCheckAction];


export const GuildConfigPatchVpnCheckAction = {
  kick: 'kick',
  ban: 'ban',
  flag: 'flag',
} as const;

export interface GuildConfigPatch {
  /** @nullable */
  logChannelId?: string | null;
  /** @nullable */
  banLogChannelId?: string | null;
  /** @nullable */
  generalLogChannelId?: string | null;
  raidMode?: boolean;
  raidMode2?: boolean;
  joinLock?: boolean;
  /** @nullable */
  ticketStaffRoleId?: string | null;
  /** @nullable */
  ticketCategoryId?: string | null;
  /** @nullable */
  transcriptChannelId?: string | null;
  welcomeEnabled?: boolean;
  /** @nullable */
  welcomeChannelId?: string | null;
  welcomeMessage?: string;
  leaveEnabled?: boolean;
  /** @nullable */
  leaveChannelId?: string | null;
  leaveMessage?: string;
  captchaEnabled?: boolean;
  /** @nullable */
  captchaChannelId?: string | null;
  /** @nullable */
  captchaUnverifiedRoleId?: string | null;
  /** @nullable */
  captchaVerifiedRoleId?: string | null;
  sanctionDmEnabled?: boolean;
  /** @nullable */
  inviteLogChannelId?: string | null;
  securityLevel?: GuildConfigPatchSecurityLevel;
  antiInsultEnabled?: boolean;
  antiInsultWords?: string[];
  antiWebhookEnabled?: boolean;
  suspiciousCheckEnabled?: boolean;
  whitelistedInviteCodes?: string[];
  vpnCheckEnabled?: boolean;
  vpnCheckMinAgeDays?: number;
  vpnCheckAction?: GuildConfigPatchVpnCheckAction;
  vpnCheckRequireNoAvatar?: boolean;
}