# lib/api-zod/src/generated/types/guildConfig.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GuildConfigSecurityLevel } from './guildConfigSecurityLevel';
import type { GuildConfigVpnCheckAction } from './guildConfigVpnCheckAction';

export interface GuildConfig {
  /** @nullable */
  logChannelId: string | null;
  /** @nullable */
  banLogChannelId: string | null;
  /** @nullable */
  generalLogChannelId: string | null;
  raidMode: boolean;
  raidMode2: boolean;
  joinLock: boolean;
  /** @nullable */
  ticketStaffRoleId: string | null;
  /** @nullable */
  ticketCategoryId: string | null;
  /** @nullable */
  transcriptChannelId: string | null;
  welcomeEnabled: boolean;
  /** @nullable */
  welcomeChannelId: string | null;
  welcomeMessage: string;
  leaveEnabled: boolean;
  /** @nullable */
  leaveChannelId: string | null;
  leaveMessage: string;
  captchaEnabled: boolean;
  /** @nullable */
  captchaChannelId: string | null;
  /** @nullable */
  captchaUnverifiedRoleId: string | null;
  /** @nullable */
  captchaVerifiedRoleId: string | null;
  sanctionDmEnabled: boolean;
  /** @nullable */
  inviteLogChannelId: string | null;
  securityLevel: GuildConfigSecurityLevel;
  antiInsultEnabled: boolean;
  antiInsultWords: string[];
  antiWebhookEnabled: boolean;
  suspiciousCheckEnabled: boolean;
  whitelistedInviteCodes: string[];
  vpnCheckEnabled?: boolean;
  vpnCheckMinAgeDays?: number;
  vpnCheckAction?: GuildConfigVpnCheckAction;
  vpnCheckRequireNoAvatar?: boolean;
}